+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed spectators sometimes spawning with health.
+- Fixed player-hurt sound inconsistency
 
 ## [v1.21.11-0.1.1] - 2026-01-17